+++ cardano-api/cardano-api.cabal
@@ -36,6 +36,7 @@ library
 
                         -- Splitting up the big Typed module:
                         Cardano.Api.Address
+                        Cardano.Api.Certificate
                         Cardano.Api.Eras
                         Cardano.Api.Error
                         Cardano.Api.Hash
